feature: clean up code and add short urls
